+++ tests/conftest.py
@@ -23,7 +23,7 @@ os.environ.pop("PINEVEX_RENDERER_ROBLOX_FONT_DIRS", None)
 BROWSER = {
     "test_billboard", "test_browser_session", "test_mcp_adapter", "test_particles",
     "test_preview", "test_preview_particles", "test_surface", "test_viewport_frame",
-    "test_visual_gallery", "test_place_realism",
+    "test_visual_gallery", "test_place_realism", "test_terrain",
 }
 NOT_LUNE = {
     "test_compare", "test_groundtruth_diff", "test_mesh_assets", "test_project",